+++ app/(white_navbar)/explore/RetroPGF1/page.tsx
@@ -49,6 +49,47 @@ export async function getJsonRound1(): Promise<RetroRound1[]> {
   return jsonData
 }
 
+export async function getSpecifigJsonRound1(
+  projectName: string
+): Promise<RetroRound1 | undefined> {
+  const directoryPath = path.join(process.cwd(), 'public/static/rpgf1.json')
+
+  const fileContents = await fs.promises.readFile(directoryPath, 'utf8')
+  const jsonData: RetroRound1[] = JSON.parse(fileContents)
+
+  jsonData.sort((a, b) => {
+    const nameA = (a['Project Name'] ?? '').toLowerCase()
+    const nameB = (b['Project Name'] ?? '').toLowerCase()
+
+    for (let i = 0; i < Math.min(nameA.length, nameB.length); i++) {
+      const charCodeA = nameA.charCodeAt(i)
+      const charCodeB = nameB.charCodeAt(i)
+
+      // Check if both characters are letters
+      const isALetter = isLetter(nameA[i])
+      const isBLetter = isLetter(nameB[i])
+
+      if (isALetter && isBLetter) {
+        // Both characters are letters, compare them normally
+        if (charCodeA !== charCodeB) {
+          return charCodeA - charCodeB
+        }
+      } else if (isALetter) {
+        // Character in nameA is a letter, prioritize it
+        return -1
+      } else if (isBLetter) {
+        // Character in nameB is a letter, prioritize it
+        return 1
+      }
+    }
+
+    // If both names are equal so far, the shorter name should come first
+    return nameA.length - nameB.length
+  })
+
+  return jsonData.find((elem) => elem['Project Name'] === projectName)
+}
+
 export default async function page() {
   const round1 = await getJsonRound1()
