+++ app/component/Statistic/RetroPGF2/AllocationEachCategory.tsx
@@ -38,10 +38,10 @@ const AllocationEachCategory: FC = () => {
         // style: {
         //   fontSize: '8px',
         // },
-        formatter: function () {
-          return this.value.length > 10
-            ? this.value.substring(0, 10) + '...'
-            : this.value
+        formatter: function (this: Highcharts.AxisLabelsFormatterContextObject) {
+          return this.value.toString().length  > 10
+            ? this.value.toString().substring(0, 10) + '...'
+            : this.value.toString();
         },
       },
